+++ packages/editor/editor-core/examples/5-full-page.tsx
@@ -216,6 +216,7 @@ export class ExampleEditorComponent extends React.Component<
                 component: ExampleInlineCommentComponent,
               }}
               allowStatus={true}
+              allowNestedTasks
               {...providers}
               media={{
                 provider: mediaProvider,