+++ packages/elements/task-decision/src/__tests__/unit/api/TaskDecisionResourceSpec.ts
@@ -2,30 +2,13 @@ import URLSearchParams from 'url-search-params';
 import fetchMock from 'fetch-mock/src/client';
 import { waitUntil } from '@atlaskit/util-common-test';
 
-import {
-  buildItemServiceResponse,
-  buildServiceDecision,
-  buildServiceTask,
-  datePlus,
-  getServiceDecisionsResponse,
-  getServiceItemsResponse,
-  getServiceTasksResponse,
-  getParticipants,
-} from '../_test-data';
-
 import TaskDecisionResource, {
   ItemStateManager,
 } from '../../../api/TaskDecisionResource';
 
-import {
-  BaseItem,
-  ObjectKey,
-  Query,
-  ServiceTask,
-  TaskState,
-} from '../../../types';
+import { ObjectKey, ServiceTask, TaskState } from '../../../types';
 
-import { objectKeyToString, toObjectKey } from '../../../type-helpers';
+import { objectKeyToString } from '../../../type-helpers';
 
 // patch URLSearchParams API for jsdom tests
 declare var global: any;
@@ -38,273 +21,6 @@ const getItemStateManager = (
 ): ItemStateManager => (resource as any).itemStateManager;
 
 describe('TaskDecisionResource', () => {
-  describe('getDecisions', () => {
-    afterEach(() => {
-      fetchMock.restore();
-    });
-
-    it('successful', () => {
-      const response = getServiceDecisionsResponse();
-      fetchMock.mock({
-        matcher: `begin:${url}`,
-        response,
-        name: 'decision',
-      });
-      const resource = new TaskDecisionResource({ url });
-      const query = {
-        containerAri: 'container1',
-        limit: 10,
-        cursor: 'cursor1',
-      };
-      return resource.getDecisions(query).then(result => {
-        const { decisions, nextQuery } = result;
-        expect(decisions.length).toBe(response.decisions.length);
-        expect(nextQuery).toBeDefined();
-        if (nextQuery) {
-          expect(nextQuery.containerAri).toEqual('container1');
-          expect(nextQuery.limit).toEqual(10);
-          expect(nextQuery.cursor).toEqual(response.meta.cursor);
-        }
-
-        const calls = fetchMock.calls('decision');
-        expect(calls.length).toBe(1);
-        const lastCallUrl = calls[0][0];
-        const options = calls[0][1];
-        expect(lastCallUrl).toEqual(`${url}decisions/query`);
-        expect(options.method).toEqual('POST');
-        const body = JSON.parse(options.body);
-        expect(body.containerAri).toEqual('container1');
-        expect(body.limit).toEqual(10);
-        expect(body.cursor).toEqual('cursor1');
-      });
-    });
-
-    it('error', () => {
-      fetchMock.mock({
-        matcher: `begin:${url}`,
-        response: 404,
-        name: 'decision',
-      });
-      const resource = new TaskDecisionResource({ url });
-      const query = {
-        containerAri: 'container1',
-        limit: 10,
-        cursor: 'cursor1',
-      };
-      return resource
-        .getDecisions(query)
-        .then(result => {
-          fail(
-            `getDecisions should return rejected promise:\n${JSON.stringify(
-              result,
-              undefined,
-              2,
-            )}`,
-          );
-        })
-        .catch(err => {
-          expect(err.code).toBe(404);
-        });
-    });
-  });
-
-  describe('getTasks', () => {
-    afterEach(() => {
-      fetchMock.restore();
-    });
-
-    it('successful', () => {
-      const response = getServiceTasksResponse();
-      fetchMock.mock({
-        matcher: `begin:${url}`,
-        response,
-        name: 'task',
-      });
-      const resource = new TaskDecisionResource({ url });
-      const query = {
-        containerAri: 'container1',
-        limit: 10,
-        cursor: 'cursor1',
-      };
-      return resource.getTasks(query).then(result => {
-        const { tasks, nextQuery } = result;
-        expect(tasks.length).toBe(response.tasks.length);
-        expect(nextQuery).toBeDefined();
-        if (nextQuery) {
-          expect(nextQuery.containerAri).toEqual('container1');
-          expect(nextQuery.limit).toEqual(10);
-          expect(nextQuery.cursor).toEqual(response.meta.cursor);
-        }
-
-        const calls = fetchMock.calls('task');
-        expect(calls.length).toBe(1);
-        const call = calls[0];
-        expect(call[0]).toEqual(`${url}tasks/query`);
-        expect(call[1].method).toEqual('POST');
-        const body = JSON.parse(call[1].body);
-        expect(body.containerAri).toEqual('container1');
-        expect(body.limit).toEqual(10);
-        expect(body.cursor).toEqual('cursor1');
-      });
-    });
-
-    it('error', () => {
-      fetchMock.mock({
-        matcher: `begin:${url}`,
-        response: 404,
-        name: 'task',
-      });
-      const resource = new TaskDecisionResource({ url });
-      const query = {
-        containerAri: 'container1',
-        limit: 10,
-        cursor: 'cursor1',
-      };
-      return resource
-        .getTasks(query)
-        .then(result => {
-          fail(
-            `getTasks should return rejected promise:\n${JSON.stringify(
-              result,
-              undefined,
-              2,
-            )}`,
-          );
-        })
-        .catch(err => {
-          expect(err.code).toBe(404);
-        });
-    });
-  });
-
-  describe('getItems', () => {
-    afterEach(() => {
-      fetchMock.restore();
-    });
-
-    it('successful', () => {
-      const response = getServiceItemsResponse();
-      fetchMock.mock({
-        matcher: `begin:${url}`,
-        response,
-        name: 'items',
-      });
-      const resource = new TaskDecisionResource({ url });
-      const query = {
-        containerAri: 'container1',
-        limit: 10,
-        cursor: 'cursor1',
-      };
-      return resource.getItems(query).then(result => {
-        const { items, nextQuery } = result;
-        expect(items.length).toBe(response.elements.length);
-        expect(nextQuery).toBeDefined();
-        if (nextQuery) {
-          expect(nextQuery.containerAri).toEqual('container1');
-          expect(nextQuery.limit).toEqual(10);
-          expect(nextQuery.cursor).toEqual(response.meta.cursor);
-        }
-
-        const calls = fetchMock.calls('items');
-        expect(calls.length).toBe(1);
-        const call = calls[0];
-        expect(call[0]).toEqual(`${url}elements/query`);
-        expect(call[1].method).toEqual('POST');
-        const body = JSON.parse(call[1].body);
-        expect(body.containerAri).toEqual('container1');
-        expect(body.limit).toEqual(10);
-        expect(body.cursor).toEqual('cursor1');
-        expect(body.sortCriteria).toEqual('CREATION_DATE');
-      });
-    });
-
-    it('error', () => {
-      fetchMock.mock({
-        matcher: `begin:${url}`,
-        response: 404,
-        name: 'items',
-      });
-      const resource = new TaskDecisionResource({ url });
-      const query = {
-        containerAri: 'container1',
-        limit: 10,
-        cursor: 'cursor1',
-      };
-      return resource
-        .getItems(query)
-        .then(result => {
-          fail(
-            `getItems should return rejected promise:\n${JSON.stringify(
-              result,
-              undefined,
-              2,
-            )}`,
-          );
-        })
-        .catch(err => {
-          expect(err.code).toBe(404);
-        });
-    });
-
-    it('sortCriteria - creationDate', () => {
-      const response = getServiceItemsResponse();
-      fetchMock.mock({
-        matcher: `begin:${url}`,
-        response,
-        name: 'items',
-      });
-      const resource = new TaskDecisionResource({ url });
-      const query: Query = {
-        containerAri: 'container1',
-        limit: 10,
-        cursor: 'cursor1',
-        sortCriteria: 'creationDate',
-      };
-      return resource.getItems(query).then(result => {
-        const { nextQuery } = result;
-        expect(nextQuery).toBeDefined();
-        if (nextQuery) {
-          expect(nextQuery.sortCriteria).toEqual(query.sortCriteria);
-        }
-
-        const calls = fetchMock.calls('items');
-        expect(calls.length).toBe(1);
-        const call = calls[0];
-        const body = JSON.parse(call[1].body);
-        expect(body.sortCriteria).toEqual('CREATION_DATE');
-      });
-    });
-
-    it('sortCriteria - lastUpdateDate', () => {
-      const response = getServiceItemsResponse();
-      fetchMock.mock({
-        matcher: `begin:${url}`,
-        response,
-        name: 'items',
-      });
-      const resource = new TaskDecisionResource({ url });
-      const query: Query = {
-        containerAri: 'container1',
-        limit: 10,
-        cursor: 'cursor1',
-        sortCriteria: 'lastUpdateDate',
-      };
-      return resource.getItems(query).then(result => {
-        const { nextQuery } = result;
-        expect(nextQuery).toBeDefined();
-        if (nextQuery) {
-          expect(nextQuery.sortCriteria).toEqual(query.sortCriteria);
-        }
-
-        const calls = fetchMock.calls('items');
-        expect(calls.length).toBe(1);
-        const call = calls[0];
-        const body = JSON.parse(call[1].body);
-        expect(body.sortCriteria).toEqual('LAST_UPDATE_DATE');
-      });
-    });
-  });
-
   describe('getTaskState', () => {
     afterEach(() => {
       fetchMock.restore();
@@ -313,19 +29,16 @@ describe('TaskDecisionResource', () => {
     const resource = new TaskDecisionResource({ url });
     const tasks = [
       {
-        containerAri: 'containerAri',
         objectAri: 'objectAri',
         localId: 'task-1',
         state: 'DONE',
       },
       {
-        containerAri: 'containerAri',
         objectAri: 'objectAri',
         localId: 'task-2',
         state: 'DONE',
       },
       {
-        containerAri: 'containerAri',
         objectAri: 'objectAri',
         localId: 'task-3',
         state: 'TODO',
@@ -354,7 +67,6 @@ describe('TaskDecisionResource', () => {
     const objectKey = {
       localId: 'task-1',
       objectAri: 'objectAri',
-      containerAri: 'containerAri',
     };
 
     describe('subscribe', () => {
@@ -401,12 +113,10 @@ describe('TaskDecisionResource', () => {
   describe('toggleTask', () => {
     const key1 = {
       localId: '1',
-      containerAri: 'c1',
       objectAri: 'o1',
     };
     const key2 = {
       localId: '2',
-      containerAri: 'c1',
       objectAri: 'o2',
     };
 
@@ -417,8 +127,6 @@ describe('TaskDecisionResource', () => {
       parentLocalId: '123',
       participants: [],
       position: 1,
-      rawContent: '[]',
-      contentAsFabricDocument: '[]',
       state: state || 'TODO',
       type: 'TASK',
     });
@@ -436,7 +144,7 @@ describe('TaskDecisionResource', () => {
     it('optimistic update', () => {
       fetchMock
         .mock({
-          matcher: 'end:tasks',
+          matcher: 'end:tasks/state',
           method: 'PUT',
           name: 'set-task',
           response: serviceTask(key1, 'DONE'),
@@ -500,7 +208,7 @@ describe('TaskDecisionResource', () => {
     it('two at same time update', () => {
       fetchMock
         .mock({
-          matcher: 'end:tasks',
+          matcher: 'end:tasks/state',
           method: 'PUT',
           name: 'set-task',
           response: (_url: any, options: any) => {
@@ -551,117 +259,6 @@ describe('TaskDecisionResource', () => {
           expect(latestState1).toBe('DONE');
           expect(latestState2).toBe('TODO');
         });
-    });
-  });
-
-  describe('recent updates', () => {
-    describe('items', () => {
-      afterEach(() => {
-        fetchMock.restore();
-      });
-
-      it('notified of recent updates', () => {
-        const resource = new TaskDecisionResource({ url });
-        const d1 = buildServiceDecision({
-          localId: 'd1',
-          lastUpdateDate: datePlus(4).toISOString(),
-        });
-        const t1 = buildServiceTask({
-          localId: 't1',
-          state: 'TODO',
-          lastUpdateDate: datePlus(3).toISOString(),
-        });
-        const d2 = buildServiceDecision({
-          localId: 'd2',
-          lastUpdateDate: datePlus(2).toISOString(),
-        });
-        const t2 = buildServiceTask({
-          localId: 't2',
-          state: 'DONE',
-          lastUpdateDate: datePlus(1).toISOString(),
-        });
-        const response = buildItemServiceResponse([d1, t1, d2, t2], {});
-
-        const t1update = buildServiceTask({
-          localId: 't1',
-          state: 'DONE',
-          lastUpdateDate: datePlus(5).toISOString(),
-        });
-        const stateUpdateResponse: BaseItem<TaskState>[] = [
-          {
-            ...toObjectKey(t1update),
-            state: t1update.state, // match service update
-            type: 'TASK',
-            lastUpdateDate: new Date(),
-          },
-        ];
-
-        fetchMock
-          .mock({
-            matcher: `begin:${url}elements/query`,
-            response,
-            times: 1,
-          })
-          .mock({
-            matcher: `begin:${url}tasks/state`,
-            response: stateUpdateResponse,
-          });
-
-        const idMock = jest.fn();
-        const recentUpdatesMock = jest.fn();
-        const handlerT1 = jest.fn();
-
-        resource.subscribe(toObjectKey(t1), handlerT1);
-
-        return resource
-          .getItems(
-            {
-              containerAri: 'cheese',
-              sortCriteria: 'lastUpdateDate',
-            },
-            {
-              id: idMock,
-              recentUpdates: recentUpdatesMock,
-            },
-          )
-          .then(response => {
-            expect(idMock.mock.calls.length).toBe(1);
-            expect(recentUpdatesMock.mock.calls.length).toBe(0);
-            expect(response.items.length).toBe(4);
-            const context = {
-              containerAri: 'cheese',
-              localId: 'bacon',
-            };
-            resource.notifyRecentUpdates(context);
-            expect(recentUpdatesMock.mock.calls.length).toBe(1);
-            expect(recentUpdatesMock.mock.calls[0][0]).toEqual(context);
-            return waitUntil(() => handlerT1.mock.calls.length === 1);
-          })
-          .then(() => {
-            expect(handlerT1.mock.calls.length).toBe(1);
-
-            const recentUpdatedId = idMock.mock.calls[0][0];
-            resource.unsubscribeRecentUpdates(recentUpdatedId);
-            resource.notifyRecentUpdates({
-              containerAri: 'cheese',
-            });
-            // No new callback as unsubscribed
-            expect(recentUpdatesMock.mock.calls.length).toBe(1);
-          });
-      });
-    });
-  });
-
-  describe('getCurrentUser', () => {
-    it('can return the current user passed in from the service config', () => {
-      const user = getParticipants(1)[0];
-      const resource = new TaskDecisionResource({ url, currentUser: user });
-      expect(resource.getCurrentUser()).toEqual(user);
-    });
-
-    it('returns undefined when currentUser is undefined in the service config', () => {
-      const resource = new TaskDecisionResource({ url });
-      expect(resource.getCurrentUser()).toBeUndefined();
     });
   });
 });