+++ packages/media/media-client/src/__tests__/unit/fileFetcher.spec.ts
@@ -1,6 +1,12 @@
 import { Auth, AuthProvider } from '@atlaskit/media-core';
-import { getFileStreamsCache } from '../..';
-import { ResponseFileItem, MediaFile, RECENTS_COLLECTION } from '../..';
+import {
+  ResponseFileItem,
+  MediaFile,
+  RECENTS_COLLECTION,
+  globalMediaEventEmitter,
+  MediaViewedEventPayload,
+  getFileStreamsCache,
+} from '../..';
 import * as MediaClientModule from '../..';
 import uuid from 'uuid';
 import { FileFetcherImpl, getItemsFromKeys } from '../../client/file-fetcher';
@@ -79,6 +85,10 @@ describe('FileFetcher', () => {
     return { fileFetcher, mediaStore, items, itemsResponse };
   };
 
+  beforeEach(() => {
+    jest.spyOn(globalMediaEventEmitter, 'emit');
+  });
+
   afterEach(() => {
     getFileStreamsCache().removeAll();
     jest.restoreAllMocks();
@@ -90,6 +100,47 @@ describe('FileFetcher', () => {
     describe('with normal browser', () => {
       beforeEach(() => {
         appendChild = jest.spyOn(document.body, 'appendChild');
+      });
+
+      it('should trigger media-viewed in globalMediaEventEmitter for recents collection', async () => {
+        const { fileFetcher } = setup();
+        await fileFetcher.downloadBinary(fileId, fileName, RECENTS_COLLECTION);
+
+        expect(globalMediaEventEmitter.emit).toHaveBeenCalledTimes(1);
+        expectFunctionToHaveBeenCalledWith(globalMediaEventEmitter.emit, [
+          'media-viewed',
+          {
+            fileId,
+            viewingLevel: 'download',
+            isUserCollection: true,
+          } as MediaViewedEventPayload,
+        ]);
+      });
+
+      it('should trigger media-viewed in globalMediaEventEmitter for non-recents collection', async () => {
+        const { fileFetcher } = setup();
+        await fileFetcher.downloadBinary(fileId, fileName, collectionName);
+
+        expect(globalMediaEventEmitter.emit).toHaveBeenCalledTimes(1);
+        expectFunctionToHaveBeenCalledWith(globalMediaEventEmitter.emit, [
+          'media-viewed',
+          {
+            fileId,
+            viewingLevel: 'download',
+            isUserCollection: false,
+          } as MediaViewedEventPayload,
+        ]);
+      });
+
+      it('should not trigger media-viewed in globalMediaEventEmitter if getFileBinaryURL fails', async () => {
+        const { fileFetcher, mediaStore } = setup();
+        asMock(mediaStore.getFileBinaryURL).mockRejectedValue(new Error());
+        try {
+          await fileFetcher.downloadBinary(fileId, fileName, collectionName);
+        } catch {
+          expect(globalMediaEventEmitter.emit).toHaveBeenCalledTimes(0);
+        }
+        expect.assertions(1);
       });
 
       it('should call getFileBinaryURL', () => {