+++ packages/media/media-client/src/client/file-fetcher.ts
@@ -26,6 +26,8 @@ import {
   MediaStoreCopyFileWithTokenBody,
   MediaStoreCopyFileWithTokenParams,
   mapMediaFileToFileState,
+  globalMediaEventEmitter,
+  RECENTS_COLLECTION,
 } from '..';
 import isValidId from 'uuid-validate';
 import { getMediaTypeFromUploadableFile } from '../utils/getMediaTypeFromUploadableFile';
@@ -475,6 +477,12 @@ export class FileFetcherImpl implements FileFetcher {
     document.body.appendChild(link);
     link.click();
     document.body.removeChild(link);
+
+    globalMediaEventEmitter.emit('media-viewed', {
+      fileId: id,
+      isUserCollection: collectionName === RECENTS_COLLECTION,
+      viewingLevel: 'download',
+    });
   }
 
   public async copyFile(