+++ packages/media/media-viewer/src/newgen/viewers/video.tsx
@@ -1,5 +1,10 @@
 import * as React from 'react';
-import { getArtifactUrl, MediaClient, FileState } from '@atlaskit/media-client';
+import {
+  getArtifactUrl,
+  MediaClient,
+  FileState,
+  globalMediaEventEmitter,
+} from '@atlaskit/media-client';
 import {
   CustomMediaPlayer,
   WithShowControlMethodProp,
@@ -51,6 +56,14 @@ export class VideoViewer extends BaseViewer<string, Props, State> {
     this.init(isHDActive);
   };
 
+  private onFirstPlay = () => {
+    const { item } = this.props;
+    globalMediaEventEmitter.emit('media-viewed', {
+      fileId: item.id,
+      viewingLevel: 'full',
+    });
+  };
+
   protected renderSuccessful(content: string) {
     const { isHDActive } = this.state;
     const { item, showControls, previewCount, onCanPlay, onError } = this.props;
@@ -68,6 +81,7 @@ export class VideoViewer extends BaseViewer<string, Props, State> {
           isHDAvailable={isHDAvailable(item)}
           isShortcutEnabled={true}
           onCanPlay={onCanPlay}
+          onFirstPlay={this.onFirstPlay}
           onError={onError}
         />
       </CustomVideoPlayerWrapper>