+++ typings/atlaskit.d.ts
@@ -3,6 +3,7 @@
  */
 declare module '@atlaskit/analytics';
 declare module '@atlaskit/codemod-util-shared-styles-to-theme';
+declare module '@atlaskit/comment';
 declare module '@atlaskit/datetime-picker';
 declare module '@atlaskit/dropdown-menu';
 declare module '@atlaskit/droplist';
@@ -17,7 +18,6 @@ declare module '@atlaskit/icon-file-type';
 declare module '@atlaskit/icon-file-type/*';
 declare module '@atlaskit/icon-object';
 declare module '@atlaskit/icon-object/*';
-declare module '@atlaskit/icon-priority';
 declare module '@atlaskit/icon-priority/*';
 declare module '@atlaskit/input';
 declare module '@atlaskit/item';
@@ -42,6 +42,7 @@ declare module '@atlaskit/textfield';
 declare module '@atlaskit/theme/colors';
 
 // Build
+declare module '@atlaskit/icon-build-process';
 declare module '@atlaskit/ssr';
 declare module '@atlaskit/docs';
 declare module '@atlaskit/build-utils';