+++ packages/core/lozenge/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/lozenge
 
+## 9.1.0
+
+### Minor Changes
+
+- [minor][10e0798da6](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/10e0798da6):
+
+  Adding an optional prop `testId` that will set the attribute value `data-testid`. It will help products to write better integration and end to end tests.
+
 ## 9.0.7
 
 ### Patch Changes